requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.28.44
botocore==1.31.44
Django==4.2.5
django-storages==1.13.2
jmespath==1.0.1
Pillow==10.0.0
python-dateutil==2.8.2
python-decouple==3.8
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==1.26.16
